requirements.txt
flask
python-docx
chromadb
openai
python-dotenv
transformers
torch
scipy
soundfile
librosa